+++ test/unit/resultsController.test.js
@@ -2,7 +2,7 @@ import ResultsController from '../../src/controllers/resultsController.js'
 import { describe, it, vi, expect, beforeEach } from 'vitest'
 
 describe('ResultsController', () => {
-  vi.mock('@/utils/asyncRequestApi.js')
+  vi.mock('@/services/asyncRequestApi.js')
 
   let asyncRequestApi
   let resultsController
@@ -14,7 +14,7 @@ describe('ResultsController', () => {
   }
 
   beforeEach(async () => {
-    asyncRequestApi = await import('@/utils/asyncRequestApi')
+    asyncRequestApi = await import('@/services/asyncRequestApi')
 
     resultsController = new ResultsController({
       route: '/results'
